abstract away some functions, fix some issues
